requirements.txt
# Multi-Agent Trading System Requirements

# Core LLM Framework
langchain>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0
openai>=1.0.0
anthropic>=0.39.0

# LLM Observability
langsmith>=0.1.0

# Data and Analysis
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.0
finnhub-python>=2.4.0
stockstats>=0.6.0

# HTTP and API Clients
requests>=2.31.0
urllib3>=2.0.0

# Web Search
tavily-python>=0.5.0

# Vector Database
chromadb>=0.5.0

# CLI and Display
rich>=13.0.0

# Dashboard and Visualization
streamlit>=1.28.0
plotly>=5.17.0

# Streaming Data (IG Markets)
lightstreamer-client-lib>=2.0.0

# Utilities
python-dotenv>=1.0.0
typing-extensions>=4.0.0
